update err stack logic.
