--- utils/errstack/errstack.go
@@ -1,88 +1,0 @@
-package errstack
-
-import (
-	"context"
-	"fmt"
-	"math"
-	"runtime"
-	"strings"
-
-	"github.com/openimsdk/protocol/errinfo"
-	"github.com/openimsdk/tools/errs"
-	"github.com/openimsdk/tools/log"
-	"github.com/openimsdk/tools/mw/specialerror"
-	"github.com/pkg/errors"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-)
-
-func HandleCallError(ctx context.Context, funcName string, args []any, err error) error {
-	log.ZWarn(ctx, "fn call WithDetails Response is error", formatError(err), "funcName", funcName, "args", args)
-	unwrap := errs.Unwrap(err)
-	codeErr := specialerror.ErrCode(unwrap)
-	if codeErr == nil {
-		log.ZError(ctx, "internal server error", formatError(err), "funcName", funcName, "args", args)
-		codeErr = errs.ErrInternalServer
-	}
-	code := codeErr.Code()
-	if code <= 0 || int64(code) > int64(math.MaxUint32) {
-		log.ZError(ctx, "unknown error code", formatError(err), "funcName", funcName, "args", args, "unknown code:", int64(code))
-		code = errs.ServerInternalError
-	}
-	grpcStatus := status.New(codes.Code(code), err.Error())
-	errInfo := &errinfo.ErrorInfo{Cause: err.Error()}
-	details, err := grpcStatus.WithDetails(errInfo)
-	if err != nil {
-		log.ZWarn(ctx, "fn call WithDetails Response is error", formatError(err), "funcName", funcName)
-		return errs.WrapMsg(err, "fn error in setting grpc status details", "err", err)
-	}
-	log.ZWarn(ctx, "fn call Response is error", details.Err())
-
-	return nil
-}
-
-func formatError(err error) error {
-	type stackTracer interface {
-		StackTrace() errors.StackTrace
-	}
-	if e, ok := err.(stackTracer); ok {
-		st := e.StackTrace()
-		var sb strings.Builder
-		sb.WriteString("Error: ")
-		sb.WriteString(err.Error())
-		sb.WriteString(" | Error trace: ")
-
-		var callPath []string
-		for _, f := range st {
-			pc := uintptr(f) - 1
-			fn := runtime.FuncForPC(pc)
-			if fn == nil {
-				continue
-			}
-			if strings.Contains(fn.Name(), "runtime.") {
-				continue
-			}
-			file, line := fn.FileLine(pc)
-			funcName := simplifyFuncName(fn.Name())
-			callPath = append(callPath, fmt.Sprintf("%s (%s:%d)", funcName, file, line))
-		}
-		for i := len(callPath) - 1; i >= 0; i-- {
-			if i != len(callPath)-1 {
-				sb.WriteString(" -> ")
-			}
-			sb.WriteString(callPath[i])
-		}
-		return errors.New(sb.String())
-	}
-	return err
-}
-
-func simplifyFuncName(fullFuncName string) string {
-	parts := strings.Split(fullFuncName, "/")
-	lastPart := parts[len(parts)-1]
-	parts = strings.Split(lastPart, ".")
-	if len(parts) > 1 {
-		return parts[len(parts)-1]
-	}
-	return lastPart
-}
